requirements.txt
orjson
snowflake-id
pymongo[srv]
python-dotenv
quart
hypercorn
quart-rate-limiter
Earl-ETF
websockets